# 03_API_CONTRACTS_OPENAPI.yaml
...
openapi: 3.0.3
info:
  title: Subscription Management Tool API
  description: |
    REST API for managing subscriptions, leads, quick quotes, and Zoho Books
    integration. All endpoints require authentication unless marked as public.

    v1.1 additions (15 May 2026) — addresses pre-implementation audit gaps:
      - PATCH /organizations/{id}, DELETE /organizations/{id}
      - POST /organizations/{id}/disconnect-zoho, /reconnect-zoho
      - POST /quick-quotes/{id}/push-to-zoho
      - POST /public/quote/{token}/view (view tracker)
      - GET /settings/audit-log + SettingsAuditLogEntry schema
      - GET /lead-conversions/{id} (poll async conversion)
      - GET /webhook-events, POST /webhook-events/{id}/retry
      - Typed /settings/{category} responses (no more `type: object`)
      - QuickQuote schema: GST breakdown fields, all timestamps, revision tracking
      - OrgBranding: columnar address (not flat string) + bank fields
  version: "1.1.0"
  contact:
    name: Excel Technologies
    email: hitesh@exceltechnologies.in
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://app.exceltechnologies.in/api
    description: Production (self-hosted)

# =====================================================================
# Security
# =====================================================================
security:
  - BearerAuth: []

# =====================================================================
# Tags (logical grouping)
# =====================================================================
...
paths:
...
  /leads:
    get:
      tags: [Leads]
      summary: List leads with filters
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25 } }
      responses:
        '200':
          description: Lead list
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items: { $ref: '#/components/schemas/Lead' }
                  pagination: { $ref: '#/components/schemas/Pagination' }

    post:
      tags: [Leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadCreateInput' }
      responses:
        '201':
          description: Lead created (may include duplicate warning)
          content:
            application/json:
              schema:
                type: object
                properties:
                  lead: { $ref: '#/components/schemas/Lead' }
                  duplicate_warning:
                    type: object
                    nullable: true
                    properties:
                      matched_lead_id: { type: string, format: uuid }
                      match_field: { type: string, enum: [email, company_name] }
        '422': { $ref: '#/components/responses/ValidationError' }
  /leads/{id}:
    get:
      tags: [Leads]
      summary: Get lead with quotes history
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Lead detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  lead: { $ref: '#/components/schemas/Lead' }
                  quotes:
                    type: array
                    items: { $ref: '#/components/schemas/QuickQuoteSummary' }

    patch:
      tags: [Leads]
      summary: Update lead fields
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadUpdateInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lead' }

    delete:
      tags: [Leads]
      summary: Archive (soft delete) lead
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204':
          description: Archived
...
  /domains:
    get:
      tags: [Domains]
      summary: List domains
      parameters:
        - { name: organization_id, in: query, schema: { type: string, format: uuid } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: Domain list
          content:
            application/json:
              schema:
                type: object
                properties:
                  domains:
                    type: array
                    items: { $ref: '#/components/schemas/Domain' }

    post:
      tags: [Domains]
      summary: Create domain mapping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain_name, organization_id, zoho_customer_id]
              properties:
                domain_name: { type: string, pattern: '^[a-z0-9.-]+\.[a-z]{2,}$' }
                organization_id: { type: string, format: uuid }
                zoho_customer_id: { type: string }
                notes: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Domain' }
  /domains/{id}:
    get:
      tags: [Domains]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Domain detail with subscriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain: { $ref: '#/components/schemas/Domain' }
                  subscriptions:
                    type: array
                    items: { $ref: '#/components/schemas/SubscriptionSummary' }

    patch:
      tags: [Domains]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                organization_id: { type: string, format: uuid }
                zoho_customer_id: { type: string }
                status: { type: string, enum: [active, inactive, suspended] }
                notes: { type: string }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Domain' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        has_more: { type: boolean }
...
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            fields:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }

    # -----------------------------------------------------------------
    # User
    # -----------------------------------------------------------------
...
    Lead:
      type: object
      properties:
        id: { type: string, format: uuid }
        lead_number: { type: string }
        company_name: { type: string }
        contact_name: { type: string, nullable: true }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        primary_domain: { type: string, nullable: true }
        gstin: { type: string, nullable: true }
        status:
          type: string
          enum: [New, Contacted, Quoted, Negotiating, Won, Converted, Lost, Archived]
        target_organization_id: { type: string, format: uuid, nullable: true }
        estimated_value: { type: number, nullable: true }
        estimated_close_date: { type: string, format: date, nullable: true }
        converted_to_zoho_customer_id: { type: string, nullable: true }
        converted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    LeadCreateInput:
      type: object
      required: [company_name, email]
      properties:
        company_name: { type: string, maxLength: 250 }
        contact_name: { type: string, maxLength: 200 }
        email: { type: string, format: email }
        phone: { type: string }
        designation: { type: string }
        primary_domain: { type: string }
        industry: { type: string }
        lead_source: { type: string }
        gstin: { type: string }
        pan: { type: string }
        billing_address_line1: { type: string }
        city: { type: string }
        state: { type: string }
        state_code: { type: string }
        postal_code: { type: string }
        country: { type: string, default: India }
        target_organization_id: { type: string, format: uuid }
        estimated_close_date: { type: string, format: date }
        estimated_value: { type: number }
        notes: { type: string }
    LeadUpdateInput:
      type: object
      properties:
        company_name: { type: string }
        contact_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        primary_domain: { type: string }
        status: { type: string }
        target_organization_id: { type: string, format: uuid }
        notes: { type: string }

    # -----------------------------------------------------------------
    # Domain
    # -----------------------------------------------------------------
    Domain:
      type: object
      properties:
        id: { type: string, format: uuid }
        domain_name: { type: string }
        organization_id: { type: string, format: uuid }
        zoho_customer_id: { type: string }
        zoho_customer_name: { type: string, nullable: true }
        status: { type: string, enum: [active, inactive, transferred, lost] }
        notes: { type: string, nullable: true }

    # -----------------------------------------------------------------
    # Subscription
    # -----------------------------------------------------------------
...
    SubscriptionSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        subscription_number: { type: string }
        zoho_item_name: { type: string }
        quantity: { type: number }
        end_date: { type: string, format: date }
        lifecycle_status: { type: string }

    # -----------------------------------------------------------------
    # Renewal History
    # -----------------------------------------------------------------
...
    QuickQuoteSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        quote_number: { type: string }
        status: { type: string }
        total_amount: { type: number }
        quote_date: { type: string, format: date }
        expiry_date: { type: string, format: date }
...